# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Pressluft API
  version: 0.2.0
  description: Canonical machine-readable API contract for Pressluft MVP.
servers:
  - url: /api
security:
  - sessionAuth: []
...
paths:
  /login:
    post:
      tags: [Auth]
      operationId: login
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          headers:
            Set-Cookie:
              description: HttpOnly session cookie containing `session_token`
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Logout
      responses:
        '200':
          description: Logout success
          headers:
            Set-Cookie:
              description: Cookie clearing header for `session_token`
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /sites:
    get:
      tags: [Sites]
      operationId: listSites
      summary: List sites
      responses:
        '200':
          description: Sites list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Site'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags: [Sites]
      operationId: createSite
      summary: Create site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSiteRequest'
      responses:
        '202':
          $ref: '#/components/responses/AcceptedJob'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
...
  /jobs:
    get:
      tags: [Jobs]
      operationId: listJobs
      summary: List jobs
      responses:
        '200':
          description: Jobs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobStatusResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session_token
...
  responses:
    AcceptedJob:
      description: Job accepted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/JobResponse'
    BadRequestError:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: Authentication required or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ConflictError:
      description: Resource state conflict or concurrency guard blocked operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      additionalProperties: false
    LoginResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
      additionalProperties: false
    LogoutResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
      additionalProperties: false
    CreateSiteRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
      additionalProperties: false
    Site:
      type: object
      required: [id, name, slug, status, created_at, updated_at, state_version]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [active, cloning, deploying, restoring, failed]
        primary_environment_id:
          type: [string, 'null']
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        state_version:
          type: integer
          minimum: 1
      additionalProperties: false
...
    JobResponse:
      type: object
      required: [job_id]
      properties:
        job_id:
          type: string
          format: uuid
      additionalProperties: false
    JobStatusResponse:
      type: object
      required: [id, status, attempt_count, max_attempts, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        job_type:
          type: string
        status:
          type: string
          enum: [queued, running, succeeded, failed, cancelled]
        site_id:
          type: [string, 'null']
          format: uuid
        environment_id:
          type: [string, 'null']
          format: uuid
        node_id:
          type: [string, 'null']
          format: uuid
        attempt_count:
          type: integer
          minimum: 0
        max_attempts:
          type: integer
          minimum: 1
        run_after:
          type: [string, 'null']
          format: date-time
        locked_at:
          type: [string, 'null']
          format: date-time
        locked_by:
          type: [string, 'null']
        started_at:
          type: [string, 'null']
          format: date-time
        finished_at:
          type: [string, 'null']
          format: date-time
        error_code:
          type: [string, 'null']
        error_message:
          type: [string, 'null']
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
...
    Error:
      $ref: ./schemas/error.json